Refactor: Move magic index filtering to separate file
